User Profile - Viewing your own profile (#23)
* Registration

Added method for registration and fixed a couple typos in Login

* database changes

Updated per comments

* First update

Added profile page. Currently, The update about me function is not working correctly.

* Second update to user profile.

The about me now updates through the modal. It reloads the page after the click.

The code for the function that updates about me might be sloppy. Refactoring may be needed, but it is working as of now.

Feedback needed on overall front end design of the page