Update common.js (#35)
* Update common.js

* Update common.js

i think this is working? I am getting back a value for total points based on what I have in transactions. 

I am unsure if a value will be available to you when you call this though. If you still think this requires a call back function please let me know. I think i got it working with a callback function as well but i opted for this as it is much cleaner.

* Update common.js

removed type